requirements.txt
faiss-cpu
transformers
accelerate
sentence-transformers
pandas
torch
gradio
